Практика Redux и React DND/ правки после ревью
