+++ src/services/actions/burger-constructor.js
@@ -1,0 +1,6 @@
+export const ADD_ITEM = 'ADD_ITEM';
+export const DELETE_ITEM = 'DELETE_ITEM';
+export const DELETE_ALL_ITEMS = 'DELETE_ALL_ITEMS';
+
+
+export const SORT_ITEMS = 'SORT_ITEMS';
