+++ src/services/actions/popup.js
@@ -1,0 +1,4 @@
+export const SHOW_ING_DETAILS = 'SHOW_ING_DETAILS';
+export const HIDE_ING_DETAILS = 'HIDE_ING_DETAILS';
+export const SHOW_ORDER_DETAILS = 'SHOW_ORDER_DETAILS';
+export const HIDE_ORDER_DETAILS = 'HIDE_ORDER_DETAILS';
