--- src/services/burger-constructor-context.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const BurgerConstructorContext = React.createContext(null);
