--- src/services/burger-ingredients-context.js
@@ -1,3 +1,0 @@
-import React from 'react';
-
-export const BurgerIngredientsContext = React.createContext(null);
