+++ src/services/reducers/burger-constructor.js
@@ -1,8 +1,8 @@
-export const ingInitialState = [];
+import {ADD_ITEM, DELETE_ALL_ITEMS, DELETE_ITEM, SORT_ITEMS} from "../actions/burger-constructor";
 
-export function burgerConstructorReducer(state, action) {
+export const burgerConstructorReducer = (state = [], action) => {
     switch (action.type) {
-        case 'add':
+        case ADD_ITEM:
             let bun = state.find((x, idx) => x.type === 'bun');
             if (action.payload.type === 'bun' && bun) {
                 if (action.payload._id !== bun._id) {
@@ -11,11 +11,15 @@ export function burgerConstructorReducer(state, action) {
                 return state;
             }
             return [...state, action.payload];
-        case 'delete':
+        case DELETE_ITEM:
             return state.filter((_, idx) => idx !== action.payload);
-        case 'deleteAll':
+        case DELETE_ALL_ITEMS:
             return [];
-        default:
-            throw new Error(`Wrong type of action: ${action.type}`);
+        case SORT_ITEMS:
+            return action.payload;
+        default: {
+            return state;
+        }
     }
-}
+};
+
